knowledge-backend/requirements.txt
Flask==2.2.5
Flask_Cors==4.0.0
neo4j==5.19.0
py2neo==2021.2.4
pyahocorasick==1.4.4
tqdm==4.64.0
numpy==1.24.3
pandas==1.5.3
python-dotenv